Add files via upload
This file is more or less scratch work. I was fiddling around with the idea of using a logit for this project, but decided against it in favor of the random forest because it would outperform the logit when there isn't linear separation in the feature space, which I'm pretty sure there isn't based on some Excel explorations. Saving in case I want to revisit any parts of this.